feat: minimal working example
Comeit runs without any actual rule logic.
The infrastructure can read rule configs from a yaml
file, parse it into a Rule object that has all the information,
call on the specific methods to run check on rules and
print out if the rules passed or not in a nice format.

It does not yet support dependencies.
Does not yet support user configurations.
Expects all rules to be ERROR's.
Does not take any input arguments.